+++ src/edutap/admin_auth/__init__.py
@@ -2,6 +2,7 @@
 
 from .backends.basic import BasicUser, basic_backend
 from .backends.federated import FederatedIdentityLike, federated_backend
+from .backends.trusted_header import trusted_header_backend
 from .identity import AdminIdentity
 from .permissions import MalformedPermission, Permission, PermissionMap
 from .requires import AdminAuth, SignInBackend, TenantResolver
@@ -20,4 +21,5 @@ __all__ = [
     "TenantResolver",
     "basic_backend",
     "federated_backend",
+    "trusted_header_backend",
 ]